move gallery related types to types/gallery
