+++ src/pages/gallery/index.tsx
@@ -52,7 +52,6 @@ import {
     sortFiles,
     sortFilesIntoCollections,
 } from 'utils/file';
-import { DateValue, Bbox } from 'types/search';
 import SearchBar from 'components/SearchBar';
 import SelectedFileOptions from 'components/pages/gallery/SelectedFileOptions';
 import CollectionSelector, {
@@ -100,6 +99,7 @@ import {
     CollectionType,
 } from 'types/collection';
 import { EnteFile, VISIBILITY_STATE } from 'types/file';
+import { GalleryContextType, SelectedState, Search } from 'types/gallery';
 
 export const DeadCenter = styled.div`
     flex: 1;
@@ -115,34 +115,6 @@ const AlertContainer = styled.div`
     font-size: 14px;
     text-align: center;
 `;
-
-export type SelectedState = {
-    [k: number]: boolean;
-    count: number;
-    collectionID: number;
-};
-export type SetFiles = React.Dispatch<React.SetStateAction<EnteFile[]>>;
-export type SetCollections = React.Dispatch<React.SetStateAction<Collection[]>>;
-export type SetLoading = React.Dispatch<React.SetStateAction<Boolean>>;
-export type setSearchStats = React.Dispatch<React.SetStateAction<SearchStats>>;
-
-export type Search = {
-    date?: DateValue;
-    location?: Bbox;
-    fileIndex?: number;
-};
-export interface SearchStats {
-    resultCount: number;
-    timeTaken: number;
-}
-
-type GalleryContextType = {
-    thumbs: Map<number, string>;
-    files: Map<number, string>;
-    showPlanSelectorModal: () => void;
-    setActiveCollection: (collection: number) => void;
-    syncWithRemote: (force?: boolean, silent?: boolean) => Promise<void>;
-};
 
 const defaultGalleryContext: GalleryContextType = {
     thumbs: new Map(),
