+++ src/utils/collection/index.ts
@@ -8,7 +8,7 @@ import { downloadFiles, getSelectedFiles } from 'utils/file';
 import { getLocalFiles } from 'services/fileService';
 import { EnteFile } from 'types/file';
 import { CustomError } from 'utils/common/errorUtil';
-import { SelectedState } from 'pages/gallery';
+import { SelectedState } from 'types/gallery';
 import { User } from 'types/user';
 import { getData, LS_KEYS } from 'utils/storage/localStorage';
 import { SetDialogMessage } from 'components/MessageDialog';
